Use the same type for l and size.

+++ src/audio_out/audio_pulse_out.c
@@ -342,7 +342,7 @@ static int ao_pulse_write(ao_driver_t *this_gen, int16_t *data,
                          uint32_t num_frames)
 {
   pulse_driver_t *this = (pulse_driver_t *) this_gen;
-  int size = num_frames * this->bytes_per_frame;
+  size_t size = num_frames * this->bytes_per_frame;
   int ret = 0;
   
   if ( !this->stream || !this->pa_class->context)
